backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
transformers>=4.36.2
torch>=2.2.0
faiss-cpu>=1.8.0
sentence-transformers>=2.3.0
google-generativeai>=0.4.0
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.2.2
